Change of font, padding and adjusted navbar to small screens
Futura: didn't work, put in a similar google font.
Changed padding of container so text doesn't go all the way to edges.
Added in css and js so the menu can have a drop down when on small
screens - however have to refresh page to get back to original navbar.